requirements.txt
Django==2.2.6
django-ckeditor==5.7.1
django-hitcount==1.3.0
django-js-asset==1.2.2
django-taggit==1.1.0
mysqlclient==1.4.4
Pillow==6.2.1
pytz==2019.3
sqlparse==0.3.0
uWSGI==2.0.18
